pyproject.toml
[project]
name = "export-glm-ocr-onnx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "accelerate>=1.12.0",
    "huggingface-hub>=1.4.1",
    "numpy>=2.4.2",
    "onnx>=1.20.1",
    "onnxruntime>=1.24.1",
    "onnxruntime-gpu>=1.24.1",
    "onnxruntime-tools>=1.7.0",
    "onnxscript>=0.6.0",
    "pillow>=12.1.0",
    "torch==2.10.0+cu130",
    "torchvision==0.25.0+cu130",
    "transformers>=5.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.15",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
